Battle Menu Update
Now there is a submenu for different options under the "attack" option for things like slashing, stabbing, impact, etc. Could be generalized in the future for different kinds of weapons (perhaps bows). Only downside is that now there are multiple pre tags in the HTML section, but that doesn't really matter. Also added in a title. Oh, and the game works now. That too.